requirements.txt
customtkinter
Pillow
paddlepaddle
paddleocr>=2.6.2
PyMuPDF
opencv-python
numpy
pydantic
pandas
openpyxl
llama-cpp-python
qrcode